two stem majors successfully cloned a repo
