+++ package.json
@@ -7,8 +7,8 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "bootstrap": "^4.6.0",
-    "pubnub": "^4.29.11",
-    "pubnub-react": "^2.1.0",
+    "pubnub": "^4.24.4",
+    "pubnub-react": "^1.3.1",
     "react": "^17.0.1",
     "react-bootstrap": "^1.5.1",
     "react-dom": "^17.0.1",
